“m

+++ weapp/apps/customerized_apps/powerme/m_powerme.py
@@ -52,39 +52,40 @@ class MPowerMe(resource.Resource):
 			total_participant_count = cache_data['total_participant_count']
 			print '================from cache'
 		else:
-			need_del_powerlogs_ids = []
-			power_logs = app_models.PowerLog.objects(belong_to=record_id)
-			power_member_ids = [p.power_member_id for p in power_logs]
-			member_id2subscribe = {m.id: m.is_subscribed for m in Member.objects.filter(id__in=power_member_ids)}
+			if username == 'weshop':
+				need_del_powerlogs_ids = []
+				power_logs = app_models.PowerLog.objects(belong_to=record_id)
+				power_member_ids = [p.power_member_id for p in power_logs]
+				member_id2subscribe = {m.id: m.is_subscribed for m in Member.objects.filter(id__in=power_member_ids)}
 
-			temp_participances = app_models.PowerMeParticipance.objects(belong_to=record_id, has_join=True).order_by('-power', 'created_at')
-			record_id2participances = {}
-			for pa in temp_participances:
-				belong_to = pa.belong_to
-				if not record_id2participances.has_key(belong_to):
-					record_id2participances[belong_to] = [pa]
-				else:
-					record_id2participances[belong_to].append(pa)
+				temp_participances = app_models.PowerMeParticipance.objects(belong_to=record_id, has_join=True).order_by('-power', 'created_at')
+				record_id2participances = {}
+				for pa in temp_participances:
+					belong_to = pa.belong_to
+					if not record_id2participances.has_key(belong_to):
+						record_id2participances[belong_to] = [pa]
+					else:
+						record_id2participances[belong_to].append(pa)
 
-			#统计助力值
-			need_power_logs = [p for p in power_logs if member_id2subscribe[p.power_member_id]]
-			power_log_ids = [p.id for p in need_power_logs]
-			need_power_member_ids = [p.be_powered_member_id for p in need_power_logs]
-			#计算助力值
-			need_power_member_id2power = {}
-			for m_id in need_power_member_ids:
-				if not need_power_member_id2power.has_key(m_id):
-					need_power_member_id2power[m_id] = 1
-				else:
-					need_power_member_id2power[m_id] += 1
-			for m_id in need_power_member_id2power.keys():
-				app_models.PowerMeParticipance.objects(belong_to=record_id,member_id=m_id).update(inc__power=need_power_member_id2power[m_id])
-			#更新已关注会员的助力详情记录
-			detail_power_member_ids = [p.power_member_id for p in need_power_logs]
-			app_models.PoweredDetail.objects(belong_to=record_id, power_member_id__in=detail_power_member_ids).update(set__has_powered=True)
-			need_del_powerlogs_ids += power_log_ids
-			#删除计算过的log
-			app_models.PowerLog.objects(id__in=need_del_powerlogs_ids).delete()
+				#统计助力值
+				need_power_logs = [p for p in power_logs if member_id2subscribe[p.power_member_id]]
+				power_log_ids = [p.id for p in need_power_logs]
+				need_power_member_ids = [p.be_powered_member_id for p in need_power_logs]
+				#计算助力值
+				need_power_member_id2power = {}
+				for m_id in need_power_member_ids:
+					if not need_power_member_id2power.has_key(m_id):
+						need_power_member_id2power[m_id] = 1
+					else:
+						need_power_member_id2power[m_id] += 1
+				for m_id in need_power_member_id2power.keys():
+					app_models.PowerMeParticipance.objects(belong_to=record_id,member_id=m_id).update(inc__power=need_power_member_id2power[m_id])
+				#更新已关注会员的助力详情记录
+				detail_power_member_ids = [p.power_member_id for p in need_power_logs]
+				app_models.PoweredDetail.objects(belong_to=record_id, power_member_id__in=detail_power_member_ids).update(set__has_powered=True)
+				need_del_powerlogs_ids += power_log_ids
+				#删除计算过的log
+				app_models.PowerLog.objects(id__in=need_del_powerlogs_ids).delete()
 
 
 			# 遍历log，统计助力值
